Fixes issue due to presence of `builder.toml`
The workflows for pipeline-builder have support to update `builder.toml`, but it currently has assumptions that the builder will be a separate repo. Since we have the builder and buildpack config in the same repo, this is causing issues with the pipeline-builder scripts.

This PR renames builder so that the scripts can run and update `buildpack.toml` and `package.toml`.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>